Working server with some requests 1.2

+++ src/test/java/ConnectionTest.java
@@ -16,9 +16,9 @@ public class ConnectionTest extends TestCase {
 
     public void testUp(){
         User.deleteAll();
-        User u1 = new Donor("1212451","Tom", LocalDate.of(1995,1,12),"tom@mail.com","0745678991","A",(float)80);
+        User u1 = new Donor("1212451","Tom", LocalDate.of(1995,1,12),"tom@mail.com","0745678991","A",80.0);
         u1.saveIt();
-        User u2 = new Donor("2314415","John",LocalDate.of(1978,2,3),"john@mail.com","0753111468","B",(float)150.6);
+        User u2 = new Donor("2314415","John",LocalDate.of(1978,2,3),"john@mail.com","0753111468","B",150.6);
         u2.saveIt();
         List<Donor> list = Donor.findAll();
         for(Donor u:list){
